Break sequential test extension method into two
This provides an overloaded method for Cards without pairs.

Previously, there was one only one method that included a default # of
pairs to considers. The logic required to make it work with and without
pairs though started to decrease its performance value.

Now, there are two methods - one for with pairs and one for without.
This helps make the logic easier to understand and maintain while also
increasing its performance.