Refactor Hand constructor to throw instead of return nil
This modifies an overloaded constructor to throw Errors instead.

Previously, when any of the preconditions were not met, nil was returned
to the caller. Unpackng and always having to account for optionals
required more trouble than they were worth though.

Now, the constructor is set to throw Errors instead of returning an
optional Hand. This enables the Hand to be used right away without
needing to check if it's nil.