Split multiple Runs into subclasses
This provides an extension of Run for each multiple Run type.

Previously, only the Run class existed for single and multiple
sequences. This ended up requiring the user to know the specific pairs
that each multiple Run type would have to contain.

Now a class exists for each multiple Run type. This makes creating them
far easier, and prevents potential logic errors when it comes to knowing
the number of pairs for a multiple Run.