Refactor Suit model to enum
This changes the Suit class into an enum.

Previously, it was designed as a class with a symbol enum and a color
enum. After test using it in the simulator app, it was immediately
obvious that creating a Suit wiht a color and symbol everytime was too
tedious and unecessary.

Now, the model is an enum. This saves many lines of code in tests, and
makes creating PlayingCards much simpler.