Refactor PlayingCard to throw instead of be optional
This changes the PlayingCard to be throw instead of returning an
optional.

Previously, the class returned nil if from a failable initializer if
something went wrong. The optional design method ended up being more of
a hassle than it was worth with all the unpacking.

Now, the constructor throws if an error occurs during initialization.
This makes it easier to create PlayingCards as will no longer have to do
checks everytime before using first.