Add multiple groups check to sequential methods
This adds an additional safety check for the # of pair groups allowed.

Previously, a TripleRun was able to be created with three different
ranked pairs. Even though it had the correct pair count, it should have
only allowed it to have one ranked grouping of three pairs.

Now, the process for creating multiple-run pairs (e.g., DoubleRun)
includes checking if multiple pair groups are allowed and if a given set
of Cards matches those requirements. This enables the TripleRun to only
allow one pair grouping set, while others could have multiple if needed.