Refactor Component protocols into classes
This changes the primary hiearchy model into classes.

Previously, these were designed as protocols. Due to lack of full
generic support in Swift, it became too difficult to implement and
override equality operators in sub modules.

Now, the protocols are classes. This enables the equality operators in
teh parent module to be overriden.