Refactor Compoenent title to use Element's title
This enables all sub-models to use title property in the Element class.

Previously, a title property was set at the individual Component
sub-model level. This became less and less efficient as more models were
starting to use the property, and a lot of them just needed the name of
the model as the title.

Now, the Element parent model uses the calling model's name as the
default title, else receives their custom version. This enables the
title to be auto-generated once without having to account for it at
every invidual sub-model level.