Transform Elements model into "interface"
This changes the Elements from an "abstract class" into an "interface".

Previously, the model was implemented as a protocol with an extension of
default members to simluate an "abstract class" in Swift. This ended up
causing too many issues down the hiearchy chain with overriding default
members and equality comparisons.

Now, the model is implemented as a pure "interface" (i.e., a protocol
with only method definitions). While there is hope for a true "abstract
class" in Swift someday, the many tradeoffs and workarounds to simulate
this experience will have to suffice for now.