Transform Mechanic models into "interfaces"
This changes the Mechanics from a "abstract classes" into "interfaces".

Previously, the model were implemented as protocols with extensions of
default members to simluate "abstract classes" in Swift. This ended up
causing too many issues down the hiearchy chain with overriding default
members and equality comparisons.

Now, the models are implemented as a pure "interfaces" (i.e., protocols
with only method definitions). While there is hope for a true "abstract
class" in Swift someday, the many tradeoffs and workarounds to simulate
this experience will have to suffice for now.