Fix Fifteen constructor's point sum check
This resolves the Error that was thrown for a valid set of Cards.

Previously, when a Fifteen was created with a valid set of Cards, an
invalidPoints Error was thrown. This prevented the tests that used these
Cards from being able to run.

Now, the Error is only thrown when appropriate. This enables all tests
to run and pass successfully.