Remove Straight and Royal Flush determinations
This removes the process for determining if a Flush is a stright or
royal Flush.

Previously, the Flush model determined these qualities at instantiation
time. The new design plan is to have separate models for these instead,
and make them extensions of the Flush model.

Now, a Flush is purely just a Flush of equally Suited Cards. This
enables it to be more extensible in the future.