Change MultiRun parent to HandRank
This switches the parent class from the Run class to the HandRank class.

Previously, the MultiRun extended the Run class. It ended up making more
sense to have it be its own entity though as many operations deal
directly with Pairs and the Run class was having to compensate too much
for it.

Now, it extends directly from the HandRank model. This frees up the Run
class to be used as is without unnecessarily complicating it.