Fix DoubleDoubleRun runs check
This resolves an issue where the creating a DoubleDouble run.

Previously, a valid DoubleDoubleRun failed to create due to its Runs not
being properly recognized as valid. Therefore, the tests that used them
also failed.

Now, a new helper method has been provided that enables validating the
Runs in a DoubleDoubleRun. As a result, all tests pass and a
DoubleDoubleRun can be successfully created.