Remove generic parameters from Hand model
This refactors the Hand class to its original design without generics.

Previously, a generic parameter was added with the expectation that it
would make certain operations easier to implement on subclassed levels.
This ended up complicated more matters though, and required being
specific about types when instantiating and comparing (e.g., even though
a PlayingCard is a RankedCard, an Array with a Pair of each set couldn't
be created without being very specifc about the type).

Now, the model is set to work with the top-level Card class. This
enables comparisons, collections holding same ultimate types, etc. to be
done with inference instead of requiring extra writing to be specific.