Change MultiRun Pair Card count to Pair count
This changes the parameter to accept the # of Pairs instead.

Previously, this was designed with the caller giving the # of Pair
Cards, but this ended up requiring too much responsibility of the caller
and was more complicated to test for.

Now, all the caller has to do is provide the # of Pairs that is
expecting in the MultiRun. This makes it easier to implement and test.