Return instance name as capitalized by default
This removes the capitalization parameter from the helper method.

Previously, the method could have been called with an option to get the
instance name capitalized or as is. Since there are no use cases right
now to obtain the "as is" version, the code will never be hit and
therefore unable to be tested.

Now, the method returns the capitalized version of the instance by
default. This ensures that no code paths in the protocol will not be
hit and tested for.