Remove error message from PlayingCardsHelper method
This removes an error message from the getCards helper method.

Previously, an error message was in place if retrieving PlayingCards
failed for some reason. This code path to this message would have never
been hit though, and therefore not necesseary and untestable.

Now, the method performs the same behaviors but without processing if
the message should be output to the console. This helps ensure that only
viable code paths are processed and tested for.