Refactor getUniqueCards to generic implementation
This enables the operation to be used by all Filterable Collections.

Previously, only Pair Arrays were able to use the method. The need to
use it with other Collection types and Element types became necessary
though.

Now, the method is generic enough to retrieve all the unique
sub-Elements from any type of Collection and Element type. This enables
greater flexibility in using it across multiple modules.