Fix low Ace being equal to high Ace with same Suit
This resolves an issue where a low Ace equaled a high Ace.

Previously, when a low Ace with the same Suit as a high Ace were
compared for equality, they were found to be equal. This was not
consistent with it also being true that a high Ace is greater than a
low Ace.

Now, a low Ace does not equal a high Ace, even if it has the same Suit.
This helps create consistent equality comparisons.