Enable creating Jokers without try statements
This enables creating a Joker without using the try keyword.

Previously, it was required to use a try keyword to create a Joker. This
did not make sense when using from the client side though as the only
parameter for the constructor was a color. The user should not have had
to account for an internal framework error.

Now, the user can create Jokers without try statements. This makes more
sense from the client side as there should be no errors for the user to
handle.