Enable creating PlayingCardDeck without try
This enables creating a new PlayingCardDeck without error handling code.

Previously, creating a standard PlayingCardDeck with or without Jokers
required using try statements. This did not make sense when using from
the client side as the only parameter was a boolean to include Jokers.

Now, the user can create a PlayingCardDeck without try statements. This
enables them to use the model without having to account for internal
errors that the framework should have resolved for them.