Refactor PlayingCardDeck's "includesJokers" field
This renames the property to "jokersAllowed".

Previously, this was named "includesJokers". This could have been
misinterpreted though as a Deck that allows Jokers in it, but doesn't
contain any at a given state.

Now, the property is named "jokersAllowed". This enables it to contain
or not contain Jokers without confusing whether they are allowed or not
in the Deck.