Add generic Card type parameter to Deck model
This refactors a Deck to hold a specific type of Card.

Previously, a Deck managed a collection of Cards. While this offered
great flexibility, it also forced a client to typecast a returned Card
when using a PlayingCardDeck subclass everytime a PlayingCard property
was needed, as well as made it possible to add other non-PlayingCards
to the Deck if protections were not in place.

Now, a Deck manages a specific Card type. This stills allows flexibility
to hold multiple Card types in a Deck if needed by using a common parent
Card type, but also provides protection for only holding specific Card
types as well as removes the need for clients to typecast returned
Cards.