Refactor Deck's cards data structure
This changes the data structure to an array.

Previously, the cards collection was implemented as a dictionary. This
offered a lot of advantages, but sorting actually turned out to be a
greater need than anticipated for a Deck in the client.

Now the cards are implemented as an array. This allows for easy sorting
in place, and shouldn't be a problem for time/space complexity as the
majority of Decks work with less than a 100 Cards.