Add generic Card type parameter to Hand model
This refactors a Hand to hold a specific type of Card.

Previously, a Hand managed a collection of Cards. While this offered
great flexibility, it also forced a client to typecast a returned Card
when using a PlayingCardHand subclass everytime a PlayingCard property
was needed, as well as made it possible to add other non-PlayingCards
to the Hand if protections were not in place.

Now, a Hand manages a specific Card type. This stills allows flexibility
to hold multiple Card types in a Hand if needed by using a common parent
Card type, but also provides protection for only holding specific Card
types as well as removes the need for clients to typecast returned
Cards.