Refactor Rank enum to be raw value independent
This removes the raw value dependency.

Previously, numeral ranks were determined by the rank order. This required a
dependency on the "numeral" ranks be listed in exact order.

Now, the numeral ranks are not dependent on any case order. This allows
new ranks to be added without needing to account for specific placement.